requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpegdash==0.4.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
ratelimit==2.2.1
requests==2.31.0
six==1.16.0
tidalapi==0.7.5
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
